requirements/prod.txt
asgiref==3.7.2
bleach==6.1.0
Django==4.2.7
django-environ==0.11.2
django-js-asset==2.1.0
django-mptt==0.15.0
djangorestframework==3.14.0
djangorestframework-recursive==0.1.2
drf_api_logger==1.1.15
gunicorn==21.2.0
packaging==23.2
psycopg2-binary==2.9.9
pytils==0.4.1
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
webencodings==0.5.1
